# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse:
    properties:
      error:
        type: string
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ValidationError'
        type: array
    type: object
...
  github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_sshconn_Response:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/internal_platform_sshconn.Response'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_Aliizi83_vohu_internal_platform_shared.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 40901
    - 50001
    type: integer
    x-enum-varnames:
    - ResultSuccess
    - ResultValidationError
    - ResultAuthError
    - ResultForbiddenError
    - ResultNotFoundError
    - ResultConflictError
    - ResultInternalError
  github_com_Aliizi83_vohu_internal_platform_shared.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
  internal_platform_sshconn.CreateSSHConnectionRequest:
    properties:
      host:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      privateKey:
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - host
    - name
    - privateKey
    - username
    type: object
  internal_platform_sshconn.Response:
    properties:
      createdByUserId:
        type: integer
      host:
        type: string
      id:
        type: integer
      name:
        type: string
      port:
        type: integer
      username:
        type: string
    type: object
  internal_platform_sshconn.UpdateSSHConnectionRequest:
    properties:
      host:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
      port:
        maximum: 65535
        minimum: 1
        type: integer
      privateKey:
        description: |-
          PrivateKey is optional on update — omitted means "keep the existing
          key," since there is no way to show it back to the client to
          prefill a form.
        type: string
      username:
        maxLength: 100
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for Vohu's multi-user platform — users, roles, polymorphic
    resource access, SSH connections, conversations/chat, provider keys, custom models,
    and the agent tool catalog. Every endpoint responds through shared.BaseResponse
    (see internal/platform/shared/response.go) — {result, success, resultCode, validationErrors,
    error} — regardless of success or failure.
  title: Vohu Platform API
  version: "1.0"
paths:
...
  /ssh-connections:
    get:
      description: Lists SSH connections the caller holds at least "read" access to
        — every connection, if the caller holds a wildcard "read"/"manage" grant on
        resource type "ssh_connection".
      parameters:
      - description: Page number, default 1
        in: query
        name: pageNumber
        type: integer
      - description: Page size, default 10
        in: query
        name: pageSize
        type: integer
      - description: JSON-encoded shared.DynamicFilter
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_sshconn_Response'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: List SSH connections
      tags:
      - ssh-connections
    post:
      consumes:
      - application/json
      description: Registers a new SSH connection. Private-key auth only — the connection
        is test-dialed (real handshake, no command run) before it's saved, so a bad
        host or mismatched key fails loudly here instead of the first time the agent
        tries to use it. The key is encrypted at rest and never returned by any response.
        The creator is auto-granted "manage" on the new connection. Requires wildcard
        "write" access on resource type "ssh_connection".
      parameters:
      - description: New connection
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_sshconn.CreateSSHConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/internal_platform_sshconn.Response'
              type: object
        "400":
          description: Validation error, or the connection test failed (bad host/key)
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Create an SSH connection
      tags:
      - ssh-connections
  /ssh-connections/{id}:
    delete:
      description: Deletes an SSH connection. Requires "manage" access to this specific
        connection.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Delete an SSH connection
      tags:
      - ssh-connections
    get:
      description: Returns one SSH connection by ID (never the private key). Requires
        at least "read" access to this specific connection.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/internal_platform_sshconn.Response'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Get an SSH connection
      tags:
      - ssh-connections
    put:
      consumes:
      - application/json
      description: Changes an SSH connection's fields. Omitting privateKey keeps the
        existing key (there's no way to show it back to the client to prefill a form).
        Requires "write" access to this specific connection.
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_sshconn.UpdateSSHConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/internal_platform_sshconn.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Update an SSH connection
      tags:
      - ssh-connections
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token from POST
      /auth/login or /auth/refresh.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"